Move code to rustc_lint.
